requirements.txt
Flask==2.0.3
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
python-dotenv==0.19.2
SQLAlchemy==1.4.31
Werkzeug==2.0.3
psycopg2